+++ mapbox_baselayer/choices.py
@@ -5,4 +5,3 @@ from django.utils.translation import gettext_lazy as _
 class LayerType(TextChoices):
     STYLE_URL = "mapbox", _("Style URL")
     RASTER = "raster", _("Raster tiles")
-    VECTOR = "vector", _("Vector tiles")